// src/shared/types/generated/application/CreateSmartFolderInput.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SmartFolderPredicate } from "./SmartFolderPredicate";

export type CreateSmartFolderInput = { name: string, parent_id: number | null, predicate: SmartFolderPredicate, icon: string | null, color: string | null, notes: string | null, sort_field: string | null, sort_order: string | null, };